feat: return TLS cert and key from get-certificate
Change the get-certificate output to return a list of certificates for
the given FQDN. The list is sorted by decreasing relevance, so the first
item has the higher relevance. This is a list of relevance factors, from
most important to less important:

- custom certificate with wildcard match
- custom certificate with exact name match
- internal (ACME) certificate with wildcard match
- internal certificate with exact name match

The match occurs in both the certificate subject and SAN extension.

If two certificates score the same relevance the order depends on the
underlying storage (custom certs or acme.json).

Expiration date is not considered.