feat!: new delete-certificate
Delete certificates matching the given serial number (for internal
certs), or path (for custom certs).

If acme.json is modified, Traefik is immediately restarted.

The names of the removed certificate are notified with
certificate-changed event. If they match an HTTP route's Host name, ACME
cert resolver is forcibly disabled for that route, to prevent further
ACME requests.