feat: implement set-default-certificate action
Completely override the default certificate configuration with a new
list of names. The first element of the list is used as main name.

Before issuing an ACME request, ensure that no certificate in acme.json
has the same set of names.

The new set of names is written to default configuration for a
time-boxed check. During this time frame, Traefik may expose its
self-signed certificate.