fix(set-certificate): avoid default cert conflicts
Use a process-wide lockfile to prevent conflict on new default cert
generation. To start a request we modify the _default_cert.yml file and
until it completes no other action run can access it.

This commit protects the request_new_default_certificate() function with
a process-wide lock file, that is released when the calling process
terminates.