fix(get-certificate): lets_encrypt attribute (#118)
The lets_encrypt attribute purpose is describing if a Let's Encrypt certificate
for the FQDN is actively requested.

With this commit, the attribute returns if Traefik is configured with a
router with acmeResolver or not for the given FQDN.

The previous implementation returned simply if a certificate for the
given FQDN was found or not.

Refs NethServer/dev#7669